env_specific different server for dev and prod
these settings show up in the client anyway. If you really wanted you
could find them, so there is no point in trying to hide the server name.
The only thing about this hard-coded server that does not make sense is
the hostname and that change will happen as soon as the host has been
migrated to our domain